requirements.txt
gensim==3.8.3
jupyterlab==2.2.9
jupyterlab-code-formatter==1.3.8
lightgbm==3.1.1
line-profiler==3.1.0
matplotlib==3.3.3
numpy==1.19.4
pandas==1.2.0
pickle==4.0
pyarrow==2.0.0
PyYAML==5.3.1
scipy==1.5.4
seaborn==0.11.1
sklearn==0.0
optuna==2.6.0
joblib==1.0.1
